Process historical data (#199)
* Exiting processor when processing is finished

* Removed bearer tokens

* Run video processing without creating extra process, don't wait for signal to die

* Removed unused dockerfile, modified getting boolean config

* Added whitespace

* Removed unused bash file, removed unused config, removed comments on supervisord conf

* Restored deleted log message

* Added section for running historical data, fixed other link

* Changed README example to bash code block

* Adding processor's automatic cameras setting

* Adding processor's automatical setting

+++ libs/engine_threading.py
@@ -2,8 +2,9 @@ import os
 import logging
 import time
 
+import sys
+
 from datetime import datetime
-from pathlib import Path
 from shutil import rmtree
 from threading import Thread
 from libs.cv_engine import CvEngine
@@ -11,7 +12,7 @@ from libs.cv_engine import CvEngine
 logger = logging.getLogger(__name__)
 
 
-def run_video_processing(config, pipe, sources):
+def run_video_processing(config, pipe, sources, historical_data_mode: bool = False):
     pid = os.getpid()
     logger.info(f"[{pid}] taking on {len(sources)} cameras")
     threads = []
@@ -20,21 +21,23 @@ def run_video_processing(config, pipe, sources):
         engine.start()
         threads.append(engine)
 
-    # Wait for a signal to die
-    pipe.recv()
-    logger.info(f"[{pid}] will stop cameras and die")
-    for t in threads:
-        t.stop()
+    if not historical_data_mode:
+        # Wait for a signal to die
+        pipe.recv()
 
-    for src in sources:
-        logger.info("Clean up video output")
-        playlist_path = os.path.join('/repo/data/processor/static/gstreamer/', src['id'])
-        birdseye_path = os.path.join('/repo/data/processor/static/gstreamer/', src['id'] + '-birdseye')
-        if os.path.exists(playlist_path):
-            rmtree(playlist_path)
-        if os.path.exists(birdseye_path):
-            rmtree(birdseye_path)
-    logger.info(f"[{pid}] Goodbye!")
+        logger.info(f"[{pid}] will stop cameras and die")
+        for t in threads:
+            t.stop()
+
+        for src in sources:
+            logger.info("Clean up video output")
+            playlist_path = os.path.join('/repo/data/processor/static/gstreamer/', src['id'])
+            birdseye_path = os.path.join('/repo/data/processor/static/gstreamer/', src['id'] + '-birdseye')
+            if os.path.exists(playlist_path):
+                rmtree(playlist_path)
+            if os.path.exists(birdseye_path):
+                rmtree(birdseye_path)
+        logger.info(f"[{pid}] Goodbye!")
 
 
 class EngineThread(Thread):
@@ -54,7 +57,7 @@ class EngineThread(Thread):
                     last_restart_time = datetime.now()
                     self.engine.process_video(self.source['url'])
                     if os.path.isdir(self.source['url']):
-                        # Source is a video. Stop the iteration
+                        logging.info("Finished processing")
                         break
                 except Exception as e:
                     logging.error(e, exc_info=True)
@@ -68,6 +71,7 @@ class EngineThread(Thread):
                     time.sleep(5)
                     logging.info("Restarting the video processing")
                     restarts += 1
+            sys.exit()
         except Exception as e:
             logging.error(e, exc_info=True)
             raise e
